requirements.txt
# Core Flask dependencies
flask>=3.1.2
flask-cors>=6.0.0

# HTTP Client
requests>=2.31.0

# Storage backends
boto3>=1.34.0
azure-storage-blob>=12.19.0

# Utilities
python-dotenv>=1.0.0

# Data validation and serialization
pydantic>=2.0.0

# Development and debugging (optional)
# pytest>=7.0.0
# pytest-flask>=1.3.0
